+++ changelog.md
@@ -1,8 +1,11 @@
 # Changelog
 
 ## Upcoming
+* Wither charging chance is no longer based off missing health, instead the charge-up time is lower when lower health 
+  * Increased time to charge from 1.5s to 2s. At 0 hp he can charge in 1s
 * Wither's blue skull chance now has lower chance to appear when below half health
 * Slight buff to wither drops
+* Fixed Wither not healing above half health
 
 ## Alpha 4.1.9
 * Fixed wither bobbing up and down when below half health